% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import dataset from RITA exports. Repair missing values and dates.}
\usage{
import(csv)
}
\arguments{
\item{csv}{file.}
}
\value{
tibble.
}
\description{
Import dataset from RITA exports. Repair missing values and dates.
}
